Agent/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.2.0
pymongo==4.6.1
redis==5.0.3
httpx==0.26.0
PyJWT==2.8.0
python-multipart==0.0.6
